+++ frontend/public/pictures/kristian-egelund-wmdcUQ0CJ4c-unsplash.jpg:Zone.Identifier
@@ -1,0 +1,4 @@
+[ZoneTransfer]
+ZoneId=3
+ReferrerUrl=https://unsplash.com/
+HostUrl=https://images.unsplash.com/photo-1469395446868-fb6a048d5ca3?ixlib=rb-1.2.1&q=80&fm=jpg&crop=entropy&cs=tinysrgb&dl=kristian-egelund-wmdcUQ0CJ4c-unsplash.jpg
