--- frontend/src/components/Body/Map.css
@@ -1,21 +1,0 @@
-/* .map_container {
-  position: absolute; 
-  left: 250px; 
-  bottom: -200px; 
-  width: 200px; 
-} */
-
-.panel {
-    position: absolute;
-    margin: 20px;
-    width: 25%;
-    top: 0;
-    bottom: 20%;
-    padding: 20px;
-    background-color: rgba(255, 255, 255, 0.9);
-    overflow-y: scroll;
-    font-family: sans-serif;
-    font-size: 0.8em;
-    line-height: 2em;
-    z-index: 1;
-}
